Property naming updates
