+++ Sources/MarkdownView/Modifiers/MathRenderingModifier.swift
@@ -13,7 +13,7 @@ extension View {
     /// - parameter enabled: A Boolean value that indicates whether to parse & render math expressions. The default value is true.
     nonisolated public func markdownMathRenderingEnabled(_ enabled: Bool = true) -> some View {
         transformEnvironment(\.markdownRendererConfiguration) { configuration in
-            configuration.mathRenderingConfiguration.enabled = enabled
+            configuration.math.shouldRender = enabled
             BlockDirectiveRenderers.shared.addRenderer(
                 MathDirectiveBlockRenderer(),
                 for: "math"
